Extract common bitmap font model and io
